requirements.txt
requests~=2.26.0
Flask~=2.0.1
Pillow~=8.3.2
spacy~=3.1.2